TST/REF: collect Index tests by method (#33870)
